finish login and sign up ui + connection
