--- NeuralPathways/ClientApp/components/home/login/login.vue.html
@@ -1,21 +1,0 @@
-﻿<template>
-    <div>
-        <h1>LogIn </h1>
-        <p>
-            Don't have an account?
-            <a href="/signup"> <strong>SignUp Here</strong></a>
-        </p>
-        <strong>
-            <p>
-                Email:
-                <input type="email" name="femail" id="userEmail"><br>
-            </p>
-            <p>
-                Password:
-                <input type="password" name="fpassword" id="userPassword"><br>
-            </p>
-            <button @click="loginButtonFunction">Login</button>
-        </strong>
-    </div>
-</template>
-<script src="./login.ts"></script>
