+++ NeuralPathways/ClientApp/components/navmenu/navmenu.css
@@ -1,66 +1,32 @@
-﻿.main-nav li .glyphicon {
-    margin-right: 10px;
+﻿body {
+    margin: 0;
+   // background-image: url("../../../wwwroot/images/bg.jpg");
+   background-color: black;
 }
 
-/* Highlighting rules for nav menu items */
-.main-nav li a.router-link-active,
-.main-nav li a.router-link-active:hover,
-.main-nav li a.router-link-active:focus {
-    background-color: #4189C7;
-    color: white;
-}
-
-/* Keep the nav menu independent of scrolling and on top of other items */
-.main-nav {
+ul {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    width: 20%;
+    background-color: white;
     position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    z-index: 1;
+    height: 100%;
+    overflow: auto;
 }
 
-@media (min-width: 768px) {
-    /* On small screens, convert the nav menu to a vertical sidebar */
-    .main-nav {
-        height: 100%;
-        width: calc(25% - 20px);
+li a {
+    display: block;
+    color: black;
+    padding: 70px 16px;
+    text-decoration: none;
+    text-align: center;
+    font-size: 24px;
+    font-family: Sylfaen;
+}
+
+
+    li a:hover:not(.active) {
+        background-color: #ff6a00;
+        color: white;
     }
-
-        .main-nav .navbar {
-            border-radius: 0px;
-            border-width: 0px;
-            height: 100%;
-        }
-
-        .main-nav .navbar-header {
-            float: none;
-        }
-
-        .main-nav .navbar-collapse {
-            border-top: 1px solid #444;
-            padding: 0px;
-        }
-
-        .main-nav .navbar ul {
-            float: none;
-        }
-
-        .main-nav .navbar li {
-            float: none;
-            font-size: 15px;
-            margin: 6px;
-        }
-
-            .main-nav .navbar li a {
-                padding: 10px 16px;
-                border-radius: 4px;
-            }
-
-        .main-nav .navbar a {
-            /* If a menu item's text is too long, truncate it */
-            width: 100%;
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-        }
-}
