+++ NeuralPathways/ClientApp/components/quiz/quiz.vue.html
@@ -1,3 +1,5 @@
 ﻿<template>
-    <h2>take a quiz here!</h2>
+    <div>
+        <p>take a quiz here!</p>
+        </div>
     </template>
