--- NeuralPathways/ClientApp/components/signup/signup.js
@@ -1,1 +1,0 @@
-﻿let userEmail = document.getElementById("userSignUpScreen").innerHTML;
