+++ NeuralPathways/ClientApp/components/signup/signup.ts
@@ -1,5 +1,4 @@
-﻿
-import Vue from 'vue';
+﻿import Vue from 'vue';
 import { Component } from 'vue-property-decorator';
 import $ from 'jquery';
 import { User } from '../../models/user';
@@ -15,7 +14,6 @@ export default class SignUpComponent extends Vue {
         this.user.lastName = (<HTMLInputElement>document.getElementById("lastName")).value;
         this.user.email = (<HTMLInputElement>document.getElementById("userEmail")).value;
         this.user.password = (<HTMLInputElement>document.getElementById("userPassword")).value;
-        
 
         //check which role the user selected
         if ((<HTMLInputElement>document.getElementById("studentRole")).checked) {
@@ -26,7 +24,6 @@ export default class SignUpComponent extends Vue {
             this.user.role = (<HTMLInputElement>document.getElementById("teacherRole")).value;
         }
 
-        
 
         $.ajax({
             headers: {
@@ -36,21 +33,20 @@ export default class SignUpComponent extends Vue {
             type: "POST",
             url: 'api/User/create',
             data: JSON.stringify({
-                FirstName: this.user.firstName,
-                LastName: this.user.lastName,
-                Email: this.user.email,
-                Password: this.user.password,
-                Role: this.user.role
+                firstname: this.user.firstName,
+                lastname: this.user.lastName,
+                email: this.user.email,
+                password: this.user.password,
+                role: this.user.role
             }),
             dataType: 'json',
-            success: function (response) {
-                alert("works");
+            complete: function (response) {
                 window.location.href = "/";
-            },
-            error: function (response) {
-                alert("sign up failed");
             }
+
         });
+
+
 
 
     }
